**fix(fetcher): resolve type mismatch in `fetchAndSaveImage` argument**
- Updated the `fetchAndSaveImage` method call in `ComicFetcher` to correctly pass the integer value directly instead of using `num.intValue()`.
- Ensures proper type handling and avoids potential runtime errors.

No breaking changes introduced.

+++ src/main/java/se/disabledsecurity/xkcd/fetcher/service/ComicFetcher.java
@@ -206,7 +206,7 @@ public class ComicFetcher implements ComicService {
         log.info("Processing backfill for comic {}", num);
 
         extractFileName(entity.getImg()).ifPresentOrElse(
-                fileName -> fetchAndSaveImage(num.intValue(), fileName),
+                fileName -> fetchAndSaveImage(num, fileName),
                 () -> log.warn("Comic {} has no valid image URL or filename", num)
         );
     }
